# swagger.yaml
openapi: 3.0.3
info:
  title: VBI-Learning API Documentation
  description: API documentation for VBI-Learning platform
  version: 1.0.0
servers:
  - url: https://vbi-learning-api.example.com
...
paths:
...
  /api/course/lesson/{slug}/{instructorId}:
    post:
      tags:
        - Course
      summary: Add a new lesson to a course
      operationId: addLesson
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: Slug of the course
        - name: instructorId
          in: path
          required: true
          schema:
            type: string
          description: ID of the instructor
      requestBody:
        description: Lesson data to add
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Lesson Title
                content:
                  type: string
                  example: Lesson content
                video:
                  type: string
                  example: video-file-name.mp4
      responses:
        '201':
          description: Lesson added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully uploaded and added lesson
        '400':
          description: Bad request or Add lesson failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Add lesson failed
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error

    put:
      tags:
        - Course
      summary: Update a lesson in a course
      operationId: updateLesson
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: Slug of the course
        - name: instructorId
          in: path
          required: true
          schema:
            type: string
          description: ID of the instructor
      requestBody:
        description: Lesson data to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                _id:
                  type: string
                  example: 615c83df157e024c78e7a294
                title:
                  type: string
                  example: Updated Lesson Title
                content:
                  type: string
                  example: Updated lesson content
                video:
                  type: string
                  example: updated-video-file-name.mp4
                free_preview:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Lesson updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '400':
          description: Bad request or Update Lesson Failed
          content:
            application/json:
              schema:
                type: string
                example: Update Lesson Failed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
  /api/course/{slug}/{lessonId}:
    put:
      tags:
        - Course
      summary: Update a lesson in a course
      operationId: updateLesson
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: Slug of the course
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
          description: ID of the lesson
      requestBody:
        description: Lesson data to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Updated Lesson Title
                content:
                  type: string
                  example: Updated lesson content
                video:
                  type: string
                  example: updated-video-file-name.mp4
                free_preview:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Lesson updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '400':
          description: Bad request or Update Lesson Failed
          content:
            application/json:
              schema:
                type: string
                example: Update Lesson Failed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
...
  /api/mark-completed:
    post:
      tags:
        - Course
      summary: Mark a lesson as completed
      operationId: markCompleted
      requestBody:
        description: Lesson and course data to mark as completed
        required: true
        content:
          application/json:
            example:
              courseId: 615c83df157e024c78e7a294
              lessonId: 615c83df157e024c78e7a295
      responses:
        '200':
          description: Lesson marked as completed
          content:
            application/json:
              example:
                ok: true
        '400':
          description: Bad request or Mark Lesson Completed Failed
          content:
            application/json:
              example: Mark Lesson Completed Failed
  /api/list-completed:
    post:
      tags:
        - Course
      summary: Get list of completed lessons for a course
      operationId: getCompletedLessons
      requestBody:
        description: Course data to get completed lessons list
        required: true
        content:
          application/json:
            example:
              courseId: 615c83df157e024c78e7a294
      responses:
        '200':
          description: List of completed lesson IDs
          content:
            application/json:
              example:
                - 615c83df157e024c78e7a295
        '400':
          description: Bad request or Error getting completed lesson
          content:
            application/json:
              example: Error getting completed lesson
...